+++ R/final.plot.R
@@ -1,36 +1,43 @@
-#rm(list=ls())
-#set.seed(20150605)
-#x <- list(X1=rnorm(100),X2=rt(50,8),X3=rchisq(80,2))
-#out <- overlapping::overlap(x)
-#OV <- out$OV
-final.plot <- function(x, OV = NULL ) {
+#' =========================================
+#' @title final.plot
+#' @description Graphical representation of estimated densities and overlapping area.
+#' @param x = list of numerical vectors to be compared; each vector is an element of the list, see \ref{\code{ovelap}}
+#' @param pairs = logical 
+#' @details It requires the ggplot2 package
+#' @author Massimiliano Pastore
+#' @examples 
+final.plot <- function (x, pairs = FALSE, boundaries = NULL ) 
+{
   
-    AREA <- group <- NULL
+  group <- NULL
+  
+  if (pairs) {
+    AREA <- NULL
     for (i1 in 1:(length(x) - 1)) {
       for (i2 in (i1 + 1):(length(x))) {
-        A <- data.frame(x = x[[i1]],
-                        group = names(x)[i1],
-                        k = paste(names(x)[i1],
-                                  names(x)[i2], sep = "-", collapse = ""))
-        B <- data.frame(x = x[[i2]],
-                        group = names(x)[i2],
-                        k = paste(names(x)[i1],
-                                  names(x)[i2], sep = "-", collapse = ""))
+        A <- data.frame(x = x[[i1]], group = names(x)[i1], 
+                        k = paste(names(x)[i1], names(x)[i2], sep = "-", 
+                                  collapse = ""))
+        B <- data.frame(x = x[[i2]], group = names(x)[i2], 
+                        k = paste(names(x)[i1], names(x)[i2], sep = "-", 
+                                  collapse = ""))
         AREA <- rbind(AREA, rbind(A, B))
       }
     }
-    
-    if (!is.null(OV)){
-      
-      OV <- data.frame(OV=OV,k=names(OV))
-      AREA <- merge(AREA,OV,by="k")
-      AREA$k <- paste0(AREA$k," (eta = ",round(AREA$OV,2),")")
-      
-    }
-    ggplot(AREA, aes(x = x, fill = group)) +
-      facet_wrap(~k) +
-      geom_density( alpha = .35) +
-      xlab("") + theme(legend.title = element_blank()) 
+  } else {
+    AREA <- data.frame(x=unlist(x),
+          group=rep(names(x), unlist(lapply(x, length))),
+          k = paste(names(x), collapse = "-"))
+  }
+  
+  OVplot <- ggplot(AREA, aes(x = x, fill = group, color=group)) +
+    theme_bw() + facet_wrap(~k) + 
+    geom_density(alpha = 0.35) + 
+    xlab("") + theme(legend.title = element_blank())
+  
+  if (!is.null(boundaries)) {
+    OVplot <- OVplot + geom_vline(xintercept = boundaries,lty=2)
+  }
+  
+  return(OVplot)
 }
-
-#final.plot(x,OV)